requirements.txt
python-dotenv==1.0.1
slack_bolt==1.21.2
slack_sdk==3.33.3
requests
bs4
spacy
nltk